+++ app/src/main/kotlin/at/bitfire/davdroid/ui/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import at.bitfire.davdroid.ui.composition.LocalNavController
+import at.bitfire.davdroid.ui.navigation.LocalNavController
 import at.bitfire.davdroid.ui.navigation.Routes
 import dagger.hilt.android.AndroidEntryPoint
